backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
pydantic-settings==2.6.1
python-multipart>=0.0.18
openai>=1.0.0
google-genai>=1.0.0
